Maintanance Original Deepsight components, Error Handling
- Native C++ changes are only Api Guard .cpp and .h and the added line for it in deepsight.vcxproj. If the C++ throws  an exception or an error the rhino crashes this handles that.
- .NET change is only NativeError.cs and the added line in DeepsightNet.csproj. The same reason as above on .net side.
- GH components, GridfromPointCloud and GridMesh were showing the original mesh and pointcloud with scaled by voxel size, sometimes the voxels dispplay started to shrink to single voxel.
Grid morphology never collected the input, malfunctioning component.
Grid combine and Grid slice didnt have vector grid support.
Grid Set Error handling, count validation if the two inputs points and values have the same list length.
GridLoad and save files were not made unless you chose a file with .vdb suffix, now it creates a .vdb file if you give it a directory.
ConstructVGrid minor change for the input parameter as int can hold limited amount of information and activevoxelscount can be a big voxel grid so the input is changed to long.
Gridmask  mismatch of transform handling. nothing major

+++ src/DeepSight.GH/VGrid/ConstructVGrid.cs
@@ -126,9 +126,9 @@ namespace DeepSight.GH.Components
             //FGrid mask = new FGrid("Mask", 0);
 
             VGrid new_grid = new VGrid("new_grid", new float[3] { 0, 0, 0 });
-            new_grid.SetActiveStates(X_grid.GetActiveVoxels(), Enumerable.Repeat(true, X_grid.ActiveVoxelCount).ToArray());
-            new_grid.SetActiveStates(Y_grid.GetActiveVoxels(), Enumerable.Repeat(true, Y_grid.ActiveVoxelCount).ToArray());
-            new_grid.SetActiveStates(Z_grid.GetActiveVoxels(), Enumerable.Repeat(true, Z_grid.ActiveVoxelCount).ToArray());
+            new_grid.SetActiveStates(X_grid.GetActiveVoxels(), Enumerable.Repeat(true, (int)X_grid.ActiveVoxelCount).ToArray());
+            new_grid.SetActiveStates(Y_grid.GetActiveVoxels(), Enumerable.Repeat(true, (int)Y_grid.ActiveVoxelCount).ToArray());
+            new_grid.SetActiveStates(Z_grid.GetActiveVoxels(), Enumerable.Repeat(true, (int)Z_grid.ActiveVoxelCount).ToArray());
 
             int[] active = new_grid.GetActiveVoxels();
             Vec3<float>[] values = new Vec3<float>[new_grid.ActiveVoxelCount];
